+++ arduinoCode/ParamList.hpp
@@ -6,13 +6,13 @@ class ParamList
 {
 private:
     Type* _dataArray;
-    uint  _numElem;
-    uint  _maxElem;
+    uint8_t  _numElem;
+    uint8_t  _maxElem;
 
 public:
     /* CONSTRUCTOR and DESTRUCTOR */
     ParamList(void) : _dataArray(nullptr), _numElem(0), _maxElem(0) { }
-    ParamList(const uint maxElem) : _dataArray(new Type), _numElem(0), _maxElem(maxElem) { }
+    ParamList(const uint8_t maxElem) : _dataArray(new Type), _numElem(0), _maxElem(maxElem) { }
     ~ParamList(void)
     {
         for (int i = 0; i < _numElem; i++)
@@ -24,7 +24,7 @@ public:
     /* ELEMENT MANAGEMENT */
     void addElem(const Type& elem) { if (_numElem < _maxElem) _dataArray[++_numElem] = elem; }
 
-    void deleteElem(const uint index)
+    void deleteElem(const uint8_t index)
     {
         if (index <= _numElem)
         {
@@ -35,7 +35,7 @@ public:
         }
     }
 
-    void setMaxElem(const uint maxElem) { _maxElem = maxElem; }
+    void setMaxElem(const uint8_t maxElem) { _maxElem = maxElem; }
 
     void increaseValue(const String& st, const Type increment)
     {
@@ -48,7 +48,7 @@ public:
 
     void clear(void) { for (int i = 0; i < _numElem; i++) this->deleteElem(i); }
 
-    Type getElem(const uint index) { return _dataArray[index]; }
+    Type getElem(const uint8_t index) { return _dataArray[index]; }
     
     Type getElemValue(const String& st)
     {
@@ -60,8 +60,8 @@ public:
         return nullptr;
     }
 
-    Type  operator[](const uint i)       { return _dataArray[i]; }
-    Type& operator[](const uint i) const { return _dataArray[i]; }
+    Type  operator[](const uint8_t i)       { return _dataArray[i]; }
+    Type& operator[](const uint8_t i) const { return _dataArray[i]; }
 };
 
 #endif
